fixed problem in excessive piping and a problem with heredoc

+++ src/cleanup.c
@@ -6,25 +6,12 @@
 /*   By: oandelin <oandelin@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/08 20:49:06 by oandelin          #+#    #+#             */
-/*   Updated: 2023/09/16 16:33:24 by oandelin         ###   ########.fr       */
+/*   Updated: 2023/09/18 17:10:35 by oandelin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 #include "lexer.h"
-
-void	malloc_error(t_data *data)
-{
-	ft_errormsg(MALLOC_FAIL, NULL);
-	reset_signals();
-	terminal_reset(data);
-	if (data->lexer.dmh_list != NULL)
-		free_list_dmh(data);
-	if (data->lexer.token != NULL)
-		free_list_token(data);
-	clear_data(data);
-	exit(1);
-}
 
 /**
  * @brief function that clears the main data struct before exiting
@@ -89,17 +76,34 @@ void	free_exec(t_exec *exec)
 {
 	t_exec	*temp;
 
+	if (exec->cmd == NULL)
+		free_empty_exec(exec);
+	else
+	{
+		while (exec)
+		{
+			free(exec->cmd);
+			free_array(exec->argv);
+			if (exec->has_heredoc)
+			{
+				if (exec->read_fd > 2)
+					close(exec->read_fd);
+				unlink(exec->heredoc);
+				free(exec->heredoc);
+			}
+			temp = exec->next;
+			free(exec);
+			exec = temp;
+		}
+	}
+}
+
+void	free_empty_exec(t_exec *exec)
+{
+	t_exec	*temp;
+
 	while (exec)
 	{
-		free(exec->cmd);
-		free_array(exec->argv);
-		if (exec->has_heredoc)
-		{
-			if (exec->read_fd > 2)
-				close (exec->read_fd);
-			unlink(exec->heredoc);
-			free(exec->heredoc);
-		}
 		temp = exec->next;
 		free(exec);
 		exec = temp;
